+++ src/DesignPattern/StructuralPatterns/Decorator/InvoicePrice/InvoiceDecorator.php
@@ -18,9 +18,9 @@ class InvoiceDecorator implements InvoicePriceInterface
     }
 
     /**
-     * @return int
+     * @return float
      */
-    public function price(): int
+    public function price(): float
     {
         return $this->invoicePrice->price();
     }
